+++ lib/Service/UserService.php
@@ -157,7 +157,14 @@ class UserService
                 } # Test for standard 'groups' attribute
                 else if (array_key_exists('groups', $casAttributes)) {
 
-                    $casGroups = (array)$casAttributes['groups'];
+                    if($this->config->getAppValue($this->appName, 'cas_groups_json_decode')) {
+
+                        $casGroups = json_decode($casAttributes['groups']);
+                    }
+                    else {
+
+                        $casGroups = (array)$casAttributes['groups'];
+                    }
                 }
 
                 foreach ($casGroups as $casGroup) {